example/otlp/requirements.txt
opentelemetry-sdk==0.17b0
opentelemetry-exporter-otlp==0.17b0
